Downloaded sub-traits now have a sensible filename.
Basically, the `finalize_response` function on the Trait view needed to be duplicated to the sub-Trait view. This function checks if the output is a FITS file, and if so, sets the `content-disposition` header on the HTTP response to include the correct filename.

Should not affect TraitProperty rendering, as they cannot be rendered as 'FITS'.

This is part of ASVO-935

branch: release-sami-1.0