Implementing the functionality to handle Archive intanciation and column associations with the instance.
This is a fairly complicated thing: Columns are defined on an `Archive` class, which, when intanciated, needs to associate the columns defined with this particular instance of the `Archive` class. This is done by creating copies of the `Column` objects that are stored in the instance dictionary. These copies are then associated with this instance of the class.

branch: trait-refactor