A bunch of work on imports and how they (don't) work.
Not sure if this is where I want it, but making stuff appear at higher levels is frought with difficulties associated with circular imports. I thought maybe I could work around them, but it seems not. FIDIA is a complex beast, which greatly benefits from being composed of multiple modules and packages, but ultimately, it presents as a single, unified thing, and shouldn't really be considered as it's individual components because it doesn't make sense that way (and hence the rise of the circular imports!)

branch: trait-refactor