Refactoring FIDIA dependencies.
I have decided not to fully follow PEP8's requirement of putting all imports at the top of a file.

There is a certain amount of interdependency involved in FIDIA packages. I have structured FIDIA such that major components are in their own pacakges (or in some cases, modules—e.g. `astro_object.py` and `sample.py`). This makes the code base easier to maintain as it is (hopefully!) readily apparent where to find code for a particular component from just the directory tree.

To make this structure look neat, each package should bring some of it's functions to the top of the package, so that other sub-components of the system do not need to know the structure of that package internally to find the things they need. Furthermore, FIDIA itself should have a fair number of tools available at the top level of the module for users to access readily. However, this structure also can cause various circular import errors. Therefore, I have decided to limit as much as possible imports within FIDIA between it's component packages. This can partially be achieved by providing a set of base classes that stand alone (now in `base_classes.py`), so that if all that is needed is e.g. to check types, then these base classes can be used. If it really is necessary to import from another sub-package of FIDIA, then this is to be done within functions as required. As this should not be common, the overhead should not be large or confusing.

This decision is based on PEP8 itself: "A Foolish Consistency is the Hobgoblin of Little Minds."

branch: trait-refactor