Refactored TraitProperties and how they are used.
TraitProperties now provide the descriptor that will retrieve a given column based on the trait's schema. TraitProperties also store information about the data type of the property, etc. They are defined on a Trait class, and specify what data the trait must provide (the "slots" if you will).

branch: trait-refactor