FIDIAColumn can now populate the entire column from a single data source, implemented for FITSBinaryTableColumn
Previously, it could only populate one object at at time with the `object_getter` function. Now there is also an `array_getter` function which pulls in the entire column in one go. Useful for places where the data is already together in a single variable (such as a FITS Binary Table). This is done using Panda's Series objects (so that the columns data can be indexed). This design decision needs to be documented in the documentation.

With that, it is now possible to properly implement FITSBinaryTableColumn, which has also been done in this commit, along with a test.

This should now enable progress on ASVO-1001, which needed some tabular data to work with.

branch: trait_collections_v0.4