Trait Mapping persistence partially working for some things.
Basically, I've set up SQLAlchemy, and I have it working at a basic level for TraitMappings and TraitPropertyMappings.

Unfortunately, SubTraitMappings are a bit harder, as they also need to have TraitPropertyMappings, but to work in the ORM, they will have to be in the same table as TraitMappings. I think this can be done, but will require more work.

ASVO-1005

branch: trait_mapping_db