Now working variant on using OrderingList in SQLAlchemy.
This basically just copies all of the code from OrderingList within SQLAlchemy into my own utilities file, but without the offending bits that are causing errors.

I'm skeptical this code is not buggy, but it seems to work and is part of SQLAlchemy (aside from my own modifications).

branch: ordering_1