Disabling Branch and Version Handling to reduce maintenance load.
This feature is not currently being used. There are no tests of it, and it's implementation needs work. However, changes in other parts of the codebase occasionally cause it wreak havoc on other tests.

Specifically at the moment, this handling is not DB persistence aware, and therefore fails when objects are restored from the database.

We do plan to include this in the near future, though, so I'm leaving all the code in place.

branch: sami_fidia_v0.4