Adds support for aliases for columns in ArchiveDefinitions.
This is not quite the solution I'd originally imagined, so I had to change the failing test to reflect the reality. Originally, I'd imagined that columns could be referred to using aliases whenever working with the original archive. However, given how Archives are now constructed from Archive Definitions, it seems to make more sense to have the constructor simply iterate through and replace aliases with actual column ids when the Archive is first constructed. That is what this code does.

Note this commit also adds iterator behaviour to TraitMappings so that it is possible to iterate over sub-mappings in a consistent way. I think this simplifies some code, but it may break things with poor test coverage. On a related note, SubTraits didn't seem to actually support their own sub-traits, as the relationship was missing. I've added that, but it is also untested.

ASVO-1000

branch: fidia-devel