Seem to have rereading of modified JSON TraitMappings into FIDIA working (ASVO-1073).
I've added a method, `.update_with_specification_dict` to TraitMapping (and related classes) that handles updating the mapping from a dictionary-like representation of the mapping (presumably that has been loaded from JSON).

Then I've modified the S7 ingestion code to (a) restructure the JSON from the original S7 format, (b) iterate over the list of mappings in `all_mappings` and update each with the corresponding.

Appears to be working, but not thoroughly tested yet.

branch: fidia-devel