JSON updater now also updates column definitions.
Some of the code for this was in place, but the actual column definitions were not being passed around, so they couldn't be updated. The `.update_with_specification_dict` method now recieves the column definitions as a dictionary-like object (ColumnDefinitionList). The TraitPropertyMapping updater can then actually do the update.

NOTE: There is no checking if columns are changed multiple times. This could be tracked, but will also be messy.

I found a few minor bugs in the process of checking this. Some unit tests would be really beneficial here, though I also think they will be hard to write.

ASVO-1073

branch: fidia-devel