S7 JSON reingester now working on Adam's udpated JSON file, including correct handling of 'dex' units.
VOUnit standard is not fully implemented in astropy around the logarithmic unit 'dex', and it seems the standard may also be incomplete, so there are some workarounds to check that.

This commit also shows that Adam's updated JSON creates a Mapping that is fully valid, so it can be used to create a FIDIA archive.

ASVO-1073

branch: fidia-devel