Added a column to provide access to the raw bytes of a file and a trait to handle computer image data.
`RawFileColumn` provides access to the bytes of a set of files. No intelligence here about the structure of those bytes.

`PixelImage` trait interprets a set of raw bytes as an Image using the Pillow (or PIL) library.

An example added to the ExampleArchive.

Basic tests included using test data.

ASVO-1071

branch: fidia-devel