Have made Sample, Archive and TraitPointer objects conform to the dictionary interface.
I wanted to use the `collections.abc.Mapping` mixin class, but it also has a metaclass, which conflicts (I think) with some of the metaclasses associated with SQLAlchemy. So I have basically copied that code into `fidia.utilities` but without the metaclass functionality. This saves implementing all functions for dict-like behavior on all of the classes.

Makes it easier to figure out what is available in dictionary-like access situations.

branch: fidia-devel