Fixed a nasty bug in they way the Trait classes (not instances) were initialized.
Turned out that when a superclass was initialized, the sub-classes would then skip their initialisation step because they inherited the flag showing that the initialisation had been done.

Flags now store the name of the class which has been initialized, so subclasses recognise when the super-class has been initialized.

branch: fidia-devel