Fixed a latent bug: Pandas data frames don't support data that isn't in native byte order.
This may be present elsewhere, and should probably be searched for. This probably should be done when column data types are standardised (e.g. are we always going to return Pandas Series for these kinds of things?)

Interestingly, this isn't checked at construction, so the bug only appears when trying to do certain actions with the DataFrame. See https://github.com/pandas-dev/pandas/issues/17396

branch: fidia-devel