Implemented a Data Access Layer.
This is fairly basic, and won't scale well yet, but the code works. Further steps will be to clean this up and make it more useable/scaleable.

The layer implemented stores the data in the file system as either numpy save files (.npy) or as pickled pandas Series objects.

Tests show the system is working, including a full ingestion of the ExampleArchive data, followed by deleting the original data and showing that FIDIA can still retrieve data.

branch: fidia-devel