Refactoring management of TraitPointers to make them less confusing.
There was some unnecessary complexity here that tended to muddle the code.

All TraitPointers now must be initialised with a pointer to the sub_mappings they point to. TraitPointer code now doesn't have a lot of if/else statements to handle whether the sub_mappings have been passed in or not.

ASVO-1186

branch: fidia-devel