This adds most of the functionality required by ASVO-1186 to Archives (but not Samples).
Traits now return either an array or iterator when asked for data via `_get_column_data` when the trait is not associated with a particular object.

Archives now set up TraitPointers in much the same way as AstronomicalObjects do.

Samples do not set up TraitPointers because knowing when to update them is much harder than I thought. I'll have to think a little more carefully about how to implement this.

Tests still not all passing because there are issues with the data coming out in the wrong order.

branch: fidia-devel