Fixed ordering issue by forcing `get_array` results to be in the same order as the sample/archive contents.
Note that the archive's contents order is not preserved because of the way the contents are stored in the persistence database. This could be fixed, but will be more work than the solution here.

Besides, it is worth checking that the order of things is always consistent, otherwise we could cause our users to have some very strange bugs without realising.

branch: fidia-devel