Found a problem where the index column for some S7 CSV Table files was incorrect.
S7_Name in the first CSV, "Galaxy" in all other files.

This does not affect the reading in of Adam's modified JSON file because the code does not check that the column IDs match, and since each TraitProperty still points to the correct column (albeit by a different name) all is fine.

It may be worth writing out a new JSON file, so that if there are further changes, we can recover to this point!

ASVO-1199

branch: fidia-devel