Added some things to Archive to make it work better when restored from the database.
There was an interesting scenario where `Archive.trait_mappings` was being initialised before the database had loaded all it's data. I'm not sure why. So, I've modified that property to check that there are actually mappings available from the database, and if not, it simply returns an empty dict and leaves everything uninitialised.

Also, since the trait pointers are not created until `trait_mappings` is initialised, I've added a `__getattr__` function that simply checks if the trait_mappings have been initialised, and if not, it initialises them before trying to the attribute lookup again.

branch: fidia-devel