Some minor additional fixes.
The number of dimensions is now passed to FIDIAColumn when a ColumnDefinition is associated with an Archive.

Type is now called `dtype` to reflect that the object stored there is a numpy `dtype` object.

At some point it becomes unclear what the difference is between a Column and a ColumnDefinition!

branch: fidia-devel