Draft of `DataCentral` archive that looks for data according to the ADC survey data ingestion requirements documents.
This covers the Catalog data only, and is incomplete. Problem is that the requirements document doesn't include a requirement for any Keys/ForeignKeys in the tables. However, FIDIA doesn't (yet) support this kind of thing either.

Furthermore, writting the code is a bit clunky with lots of nested for loops. I want to see if I can make this cleaner.

ASVO-1267

branch: adc-validator