Fixed an error in the setup of the relationship between columns and archives.
SQLAlchemy relationships need the back_populates if they are to refer to attributes on both sides of the relationship. Otherwise, they can get out of sync. This was causing a difficult to reproduce/debug error in the ADC Archive development.

Unfortunately, I'm not sure I know how to create a test that reproduces this, since the example archive should have shown it.

branch: master